+++ src/main/java/in/koreatech/koin/domain/user/verification/repository/UserDailyVerificationCountRedisRepository.java
@@ -1,10 +1,10 @@
-package in.koreatech.koin.domain.user.repository;
+package in.koreatech.koin.domain.user.verification.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.repository.Repository;
 
-import in.koreatech.koin.domain.user.model.UserDailyVerificationCount;
+import in.koreatech.koin.domain.user.verification.model.UserDailyVerificationCount;
 
 public interface UserDailyVerificationCountRedisRepository extends Repository<UserDailyVerificationCount, String> {
